+++ src/components/common/index.ts
@@ -1,0 +1,3 @@
+export { ProtectedRoute } from "./ProtectedRoute";
+export { LoadingSpinner } from "./LoadingSpinner";
+export { ErrorMessage } from "./ErrorMessage";